requirements.txt
wheel==0.33.6
uWSGI==2.0.18
python-dotenv==0.10.3
Flask==1.1.1
psycopg2-binary==2.8.3
marshmallow==2.20.1
pandas==0.25.0